pyproject.toml
[project]
name = "ado-base"
description = """ado is a unified platform for executing computational experiments at scale and analysing their results.
It can be easily extended with new experiments or new analysis tools. It allows distributed teams of
researchers and engineers to collaborate on projects, execute experiments, and share data.
"""
keywords = ["discovery", "exploration", "optimization", "orchestration", "ray"]
readme = "README.md"
license-files = ["LICEN[CS]E*", "vendored/licenses/*.txt", "AUTHORS.md"]
authors = [
    { name = "Michael A. Johnston", email = "michaelj@ie.ibm.com" }
]
maintainers = [
    { name = "Michael A. Johnston", email = "michaelj@ie.ibm.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Deps/Build-related section
dynamic = ["version"]
requires-python = ">=3.10,<3.13"
dependencies = [
    "build>=1.3.0",
    "colorlog>=6.9.0",
    "jsonpath-ng>=1.7.0",
    "jupyter>=1.1.1",
    "numpy<2",
    "pandas>=2.2.0",
    "py-markdown-table>=1.3.0",
    "pydantic>=2.10",
    "pymysql[rsa]>=1.1.1",
    "pyyaml>=6.0.2",
    "ray[serve]>=2.9",
    "scipy>=1.15.3",
    "sqlalchemy>2",
    "typer>=0.16.1",
]

[project.urls]
Homepage = "https://pages.github.ibm.com/Discovery-Orchestrator/ad-orchestrator/"
#Documentation = "https://readthedocs.org"
Repository = "https://github.ibm.com/Discovery-Orchestrator/ad-orchestrator.git"
Issues = "https://github.ibm.com/Discovery-Orchestrator/ad-orchestrator/issues"
Changelog = "https://github.ibm.com/Discovery-Orchestrator/ad-orchestrator/CHANGELOG.md"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.10.2",
    "pip>=25.1.1",
    "pre-commit>=4.2.0",
    "pydevd>=3.3.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-recording>=0.13.4",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.12.1",
    "testcontainers>=4.10.0",
    "tox>=4.26.0",
    "tox-uv>=1.25.0",
]
docs = [
    "htmlmin>=0.1.12",
    "mkdocs>=1.6.1",
    "mkdocs-github-admonitions-plugin>=0.1.1",
    "mkdocs-include-markdown-plugin>=7.1.6",
    "mkdocs-material>=9.6.14",
    "mkdocs-minify-plugin>=0.8.0",
    "pygments>=2.19.1",
    "pymdown-extensions>=10.15",
]
resolution-helpers = [
    "urllib3>=2.5.0",
]

[tool.uv]
default-groups = ["dev", "docs"]

[project.scripts]
ado = "orchestrator.cli.core.cli:app"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
pattern = "default-unprefixed"
fallback-version = "0.0.0"
tagged-metadata = true
dirty = true
bump = true
